requirements.txt
flask
flask-cors
firebase-admin
python-dotenv
flasgger
pyJWT